+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Changes
 * [UI]: Updated clients listing table to use `ant.design` tables.
 * [UI]: Sending email to sync user when a remote project's token expires.
 * [UI]: Fixed bug that prevented server side filtering for analysis name on analyses table. (20.01.01)
+* [UI]: Updated admin announcements table and creation to use `ant.design`.
+* [UI]: Fixed bug that preventing using the `ngs-linker` script with >= 330 samples.
+* [UI]: Updated breadcrumbs to have cleaner look and feel and include the actual project and sample names.
+* [UI/Developer]: Updated to `react` 16.13.0.
 * [API/Developer]: Changed the `SequencingRun` type to allow a user-defined `sequencerType` instead of it being hardcoded to `MiSeq`.
 
 19.09 to 20.01